add foreground listener to update the database when the app is open. If there is no internet connection the database is gona keep the old data

+++ app/src/main/java/com/rhodar/mobile/codescrum/validtest/data/responseModels/ArtistData.kt
@@ -4,7 +4,7 @@ import com.google.gson.annotations.SerializedName
 
 data class ArtistData (
         val name: String,
-        val listeners : Int,
+        val listeners : Long,
         val mbid: String,
         val url : String,
         val streamable : Int,
